requirements.txt
upsonic>=0.69.1
httpx>=0.26.0
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.31.1
